feat: rename resources to features
