+++ src/Agenda.API/Features/Appointments/AttendeeInfo.cs
@@ -1,0 +1,24 @@
+﻿using Agenda.Ids;
+
+namespace Agenda.API.Features.v1.Appointments;
+
+/// <summary>
+/// A person who participate to an appointment
+/// </summary>
+public record AttendeeInfo : Resource<AttendeeId>
+{
+    /// <summary>
+    /// Name of the participant
+    /// </summary>
+    public string Name { get; init; }
+
+    /// <summary>
+    /// Phone number
+    /// </summary>
+    public string PhoneNumber { get; init; }
+
+    /// <summary>
+    /// Email
+    /// </summary>
+    public string Email { get; init; }
+}
