+++ src/Agenda.API/Features/Appointments/v1/Delete/DeleteByIdRequest.cs
@@ -1,0 +1,9 @@
+using Agenda.Ids;
+
+namespace Agenda.API.Features.Appointments.v1.Delete;
+
+/// <summary>
+/// Request to delete an appointment by its identifier
+/// </summary>
+/// <param name="Id">Identifier of the appointment to delete</param>
+public record DeleteByIdRequest(AppointmentId Id);
