--- src/Agenda.API/Resources/Appointments/AttendeeInfo.cs
@@ -1,24 +1,0 @@
-﻿using Agenda.Ids;
-
-namespace Agenda.API.Resources.v1.Appointments;
-
-/// <summary>
-/// A person who participate to an appointment
-/// </summary>
-public record AttendeeInfo : Resource<AttendeeId>
-{
-    /// <summary>
-    /// Name of the participant
-    /// </summary>
-    public string Name { get; init; }
-
-    /// <summary>
-    /// Phone number
-    /// </summary>
-    public string PhoneNumber { get; init; }
-
-    /// <summary>
-    /// Email
-    /// </summary>
-    public string Email { get; init; }
-}
