--- src/Agenda.API/Resources/Appointments/v1/Delete/DeleteByIdRequest.cs
@@ -1,9 +1,0 @@
-using Agenda.Ids;
-
-namespace Agenda.API.Resources.Appointments.v1.Delete;
-
-/// <summary>
-/// Request to delete an appointment by its identifier
-/// </summary>
-/// <param name="Id">Identifier of the appointment to delete</param>
-public record DeleteByIdRequest(AppointmentId Id);
