+++ tests/Agenda.API.IntegrationTests/Appointments/v1/Create/CreateAppointmentEndpointShould.cs
@@ -7,9 +7,9 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using Agenda.API.IntegrationTests.Fixtures;
-using Agenda.API.Resources;
-using Agenda.API.Resources.Appointments;
-using Agenda.API.Resources.v1.Appointments;
+using Agenda.API.Features;
+using Agenda.API.Features.Appointments;
+using Agenda.API.Features.v1.Appointments;
 using Agenda.Ids;
 using Aspire.Hosting;
 using Aspire.Hosting.Testing;
